AquaTrack/requirements.txt
pydantic-settings~=2.11.0
SQLAlchemy~=2.0.43
fastapi~=0.118.0
pydantic~=2.11.9
passlib~=1.7.4
python-jose~=3.5.0
starlette~=0.48.0